First gc allocation!
Instead of vm owning a heap, create a heap in main interprete call and
pass the reference to it to both the compiler and vm.

Compiler parses string literal and allocates StrObj on the heap.

Implement `sweep` in Heap which walks through the linked list of
objects and deallocates unmarked objects. Call the sweep while
dropping the Heap to clean up the allocations before exiting or
starting a new repl prompt.