Support jumps during execution; if statement implemented
Refactored chunk to expose apis to read instructions directly using an
index into the code chunk. Vm no longer uses iterator to execute the
code chunk, instead, it uses ip as a direct index into the code chunk
which allows it to do code jumps easily.

Since vm no longer has the idea of instruction count, the debug line
numbers need to be accessed using ip now. So change the lines vector
encoding to be one-to-one mapping with the code vector elements.

Debugger still uses iterator to read and print the code chunk.