Implemented full support for closures
In compiler, added new type UpValue which represents a closure
capture. FunState stores a vector of such upvalues. As part of
resolving a variable name, compiler adds a new upvalue for each
capture detected (if name is resolved as local variable of enclosing
function or upvalue in the enclosing function). At the end of function
declaration, compiler emits new OpCapture* instructions right after
OpClosure which tell vm to capture variables in the newly created
closure. Compiler also keep track of which local has been captured,
and when that local gets destroyed (as part of going out of scope),
compiler emits new OpCloseUpval instruction, which tells vm to hoist
the local from stack onto the heap.

In VM, implement new instructions which manages lifetime of
captures.

Fixed main to interprete a src file and added a few toy examples.